+++ encoder_config_test.go
@@ -46,7 +46,7 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
 							"file.line": 30,
 							"file.name": "ecszap/json_encoder_test.go"
 						},
-						"log.origin.stacktrace": "stacktrace frames",
+						"log.origin.stack_trace": "frames",
 						"log.logger": "ECS",
 						"foo": "bar",
 						"dur": 5000000}`},
@@ -58,7 +58,7 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
 						"dur": 5000000}`},
 		{name: "allEnabled",
 			input: `{"enableName": true, 
-  					 "enableStacktrace": true,
+  					 "enableStackTrace": true,
 					 "enableCaller":true,
 					 "levelEncoder": "upper",
 				 	 "durationEncoder": "ms",
@@ -70,7 +70,7 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
 							"file.line": 30,
 							"file.name": "ecszap/json_encoder_test.go"
 						},
-						"log.origin.stacktrace": "stacktrace frames",
+						"log.origin.stack_trace": "frames",
 						"log.logger": "ECS",
 						"foo": "bar",
 						"dur": 5}`},
@@ -93,7 +93,7 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
 				Time:       time.Unix(1599994083, 0).UTC(),
 				Message:    "log message",
 				Caller:     caller,
-				Stack:      "stacktrace frames",
+				Stack:      "frames",
 				LoggerName: "ECS",
 			}
 			fields := []zapcore.Field{
@@ -134,7 +134,7 @@ func TestECSCompatibleEncoderConfig(t *testing.T) {
 			cfg: zapcore.EncoderConfig{
 				MessageKey: "replaced messageKey", LevelKey: "replaced levelKey",
 				TimeKey: "replaced timeKey", EncodeTime: zapcore.EpochMillisTimeEncoder,
-				NameKey: "replaced nameKey", StacktraceKey: "replaced stacktraceKey",
+				NameKey: "replaced nameKey", StacktraceKey: "replaced stackTraceKey",
 				CallerKey: "replaced callerKey", EncodeLevel: zapcore.CapitalLevelEncoder},
 			expected: `{"log.level": "DEBUG",
 						"@timestamp": "2020-09-13T10:48:03.000Z",
@@ -143,7 +143,7 @@ func TestECSCompatibleEncoderConfig(t *testing.T) {
 							"file.line": 30,
 							"file.name": "ecszap/json_encoder_test.go"
 						},
-						"log.origin.stacktrace": "stacktrace frames",
+						"log.origin.stack_trace": "frames",
 						"log.logger": "ECS",
 						"foo": "bar",
 						"count": 8}`},
@@ -155,7 +155,7 @@ func TestECSCompatibleEncoderConfig(t *testing.T) {
 				Time:       time.Unix(1599994083, 0).UTC(),
 				Message:    "log message",
 				Caller:     caller,
-				Stack:      "stacktrace frames",
+				Stack:      "frames",
 				LoggerName: "ECS",
 			}
 			fields := []zapcore.Field{